Added in exception handling for forbidden and unauthorized exceptions
