+++ client/translations/error-pages.en.yml
@@ -4,3 +4,8 @@ errors:
       htmlTitle: "503 - Service currently unavailable"
       pageTitle: "Service currently unavailable (503 error)"
       pageSectionDescription: "Try refreshing the page or returning later"
+  access-denied:
+    page:
+      htmlTitle: "Access Denied"
+      pageTitle: "Access Denied"
+      pageSectionDescription: "You do not have permission to view this page."
